Adding utility code for measuring CPU usage
I have noticed that during execution of one of 'key' tests the CPU usage goes to 100%. In order to keep an eye on CPU usage during test execution, I have created a code that measures the usage in the background. I just have to somehow include the measurement in test's pass/fail evaluation.