Improvements to test framework
Some libcw test framework functions were accepting "errors_only" argument that most of the time was set to false (actually to zero, because it was easier for me to toggle 0/1 than to toggle true/false). However this rarely used argument made the code less readable, so I decided to provide two alternative versions of functions that don't have the argument, but behave as if the argument was set to true or to false.
Also adding -Werror to CFLAGS for unit tests.