Removing concept of 'tone queue key'
State of generator is now tracked by 'state_tracking' variable in generator. It is clearer and more explicit to use this variable to keep track of state of generated tones than to use a key. Key has clearly defined purpose, and stretching it to keep track of generator/tone queue state was... stretching it.

For now registration of 'state callback' function is done per generator, not per key.