Initial commit of new test for 'state change' callback
The test verifies that calls to registered 'state change' callback are made at more-or-less expected intervals.
This test has been initially thought as test for correct behaviour of ALSA buffer. The behaviour has been fixed some time ago in commit cb99e7884dc5370519dc3a3eacfb3184959b0f87. But then I thought that this test can and should be called for other sound systems as well. So this is a test executed for 'gen' module and all supported sound systems.