Refactoring 'state change' callbacks code
Removing duplication of 'state change' callback code. It existed in key and in generator, now it exists only in generator. Not saying that this is the best solution: it may be changed in the future. But for now we have a single place where the callback is being called: in generator.

Simplifying prototype of callback function.