Per-sound-system sound device in tests
It is now possible to pass to test code a name of sound device to be used with particular sound system.
For now only ALSA sound device can be specified, but it should be very easy to extend the code to OSS device or console device.

The libcw test code can be executed with multiple sound systems. Until now it was impossible to say "use default device for Console and OSS, but use this device for ALSA". It is now possible to do so.