Undoing some changes made between release 3.5.1 and now
Undoing some changes made in non-library parts of the project. There
are already lots of changes in libcw, I don't want to make the diff
between previous release and current release unnecessarily large.

This is not very professional, and I'm losing some work, but it will
be easier for me to control and review changes between releases.