Refactoring usage of common receiver code
Some receiver code that is similar in xcwcp and libcw's tests is now moved to
cwutils/cw_rec_utils* files. Both places share some data structures and
functions.

It will take some work to make the two places share more code, and I think
that it may be impossible to have all code shared in that way. But at least
some of it can be shared.