Misc improvements
Misc improvements to code done while working on a test code for cwdaemon.

Most notable improvement is adding cw_easy_receiver_poll(): a wrapper for
polling of spaces or characters. The code that calls polling space or
character should look more or less the same in all applications. The wrapper
can be used in a simple receiver code, and more advanced functions can follow
the flow of logic in that wrapper.